using pdl's interpolation to seek in text to audio
uses hotkey combo "<a>s" Mnemonic: <a>udio Seek

PDL's a helluva dependency for just that, but the interpolation
is very primative and I expect that PDL could improve greatly
on it.

interpolation points are left in text & audio every time a label is
generated. It's not exact, and some of them will likely be totally
wrong. PDL should filter out the ones that are obviously wrong and it
should use a bunch more everywhere. Deletion undos should recover their
position as well.

The next task is obviously seeking in audio to text pos.